refactor(web): remove unnecessary barrel re-export files
Delete 5 barrel index.ts files and update all consumers to import
directly from source modules for clearer dependency graphs.
