--- apps/web/src/app/(main)/(explore)/cars/components/category/index.ts
@@ -1,11 +1,0 @@
-export { CategoryHeroCard } from "./category-hero-card";
-export { CategoryInsightsCard } from "./category-insights-card";
-export { type CategoryConfig, CategoryOverview } from "./category-overview";
-export {
-  CategoryInsightsCardSkeleton,
-  CategorySummaryCardSkeleton,
-  CategoryTabsPanelSkeleton,
-} from "./category-skeletons";
-export { CategorySummaryCard } from "./category-summary-card";
-export { CategoryTabsPanel } from "./category-tabs-panel";
-export { TopMakesChart } from "./top-makes-chart";
