--- apps/web/src/app/(main)/(explore)/cars/components/makes/index.ts
@@ -1,11 +1,0 @@
-export { AllMakes } from "./all-makes";
-export { CoeComparisonChart } from "./coe-comparison-chart";
-export { MakeCard } from "./make-card";
-export { MakeDetail } from "./make-detail";
-export { MakeGrid } from "./make-grid";
-export { MakeSearch } from "./make-search";
-export { MakeTrendChart } from "./make-trend-chart";
-export { MakesDashboard } from "./makes-dashboard";
-export { MakesSummary } from "./makes-summary";
-export { PopularMakes } from "./popular-makes";
-export { TypeBreakdownChart } from "./type-breakdown-chart";
