--- apps/web/src/app/(main)/(explore)/coe/components/pqp/index.ts
@@ -1,6 +1,0 @@
-export { ComparisonMixedChart } from "./comparison-mixed-chart";
-export { ComparisonSummaryCard } from "./comparison-summary-card";
-export { DataTable } from "./data-table";
-export { InfoPopover } from "./info-popover";
-export { RenewalCalculator } from "./renewal-calculator";
-export { TrendsChart } from "./trends-chart";
