+++ apps/web/src/queries/cars/index.ts
@@ -4,7 +4,12 @@ export * from "./compare";
 export * from "./electric-vehicles";
 export * from "./filter-options";
 export * from "./latest-month";
-export * from "./makes";
+export * from "./makes/current-year-popular-makes";
+export * from "./makes/entity-breakdowns";
+export * from "./makes/entity-checks";
+export * from "./makes/get-make-from-slug";
+export * from "./makes/grouped-makes";
+export * from "./makes/registration-stats";
 export * from "./market-insights";
 export * from "./monthly-registrations";
 export * from "./yearly-statistics";
