--- apps/web/src/queries/cars/makes/index.ts
@@ -1,6 +1,0 @@
-export * from "./current-year-popular-makes";
-export * from "./entity-breakdowns";
-export * from "./entity-checks";
-export * from "./get-make-from-slug";
-export * from "./grouped-makes";
-export * from "./registration-stats";
