--- apps/web/src/queries/index.ts
@@ -1,4 +1,0 @@
-export * from "./cars";
-export * from "./coe";
-export * from "./logos";
-export * from "./vehicle-population";
